+++ CHANGELOG.md
@@ -17,6 +17,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - _Experimental_: Add `user-valid` and `user-invalid` variants ([#12370](https://github.com/tailwindlabs/tailwindcss/pull/12370))
 - _Experimental_: Add `wrap-anywhere`, `wrap-break-word`, and `wrap-normal` utilities ([#12128](https://github.com/tailwindlabs/tailwindcss/pull/12128))
 
+## Fixed
+
+- Ensure that all CSS variables used by preflight are prefixed ([#17036](https://github.com/tailwindlabs/tailwindcss/pull/17036))
+
+### Changed
+
+- The `--theme(…)` function now returns CSS variables unless used inside positions where CSS variables are invalid (e.g. `@media` queries) ([#17036](https://github.com/tailwindlabs/tailwindcss/pull/17036))
+
 ## [4.0.12] - 2025-03-07
 
 ### Fixed